requirements.txt
# requirements.txt
# 长线 A 股数据分析项目依赖

# === 数据源 ===
tushare>=1.4.0

# === 数据处理 ===
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0          # Parquet 读写

# === 金融计算 ===
scipy>=1.11.0
statsmodels>=0.14.0

# === 回测框架 ===
backtrader>=1.9.78

# === 可视化 ===
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# === 报告生成 ===
jinja2>=3.1.0
weasyprint>=60.0         # PDF 导出 (可选)

# === 配置与工具 ===
pyyaml>=6.0
python-dateutil>=2.8.0
tqdm>=4.65.0

# === 日志 ===
loguru>=0.7.0
